requirements.txt
PyQt5>=5.15.0
opencv-python>=4.7.0
numpy>=1.21.0
pillow>=9.0.0
pygetwindow>=0.0.9
pyautogui>=0.9.53
websockets>=10.0
